requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pyautogen>=0.2.0
